requirements.txt
pandas==1.5.3
tqdm==4.64.1
numpy==1.23.2
torch==2.3.1
torchvision==0.18.1
timm==1.0.7
faiss-cpu==1.7.4
requests==2.28.1
Pillow==9.3.0